Implementing user REST APIs

+++ src/main/java/com/example/e_commerce/model/Category.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Category {
-    private Long category_id;
+    private int category_id;
     private String category_name;
     private String description;
 }
